Fix bug for $ in WiFi-passwords
This fix aims to solve incompatibility of $ inside Wifi-Passwords. I have narrowed down the errors we have experienced with one particular WiFi network to stem from this line in the code.

I was considering solving this by overriding the SetWiFi global variable for our usecase, but this would not be feasible as the SetWifi variable is defined and updated inside the main OSD script based upon whether or not WiFi is connected, and is likely to cause issues elsewhere.