Update Start-OSDCloud.ps1
The ImageFileItem parameter has been added, which updates the ImageFileItem property in the $global:StartOSDCloud variable. If ImageFileItem is provided, the existence of the specified file is checked. If the file does not exist, the user will be prompted to manually select an ImageFileItem as usual.

The integer value for OSImageIndex is also validated based on the provided ImageFileItem. If the specified index does not exist, the user will be prompted to manually select an item, as per the usual process.

In Summary: 
It is now possible to perform a zero-touch deployment for offline images.

Example:
Start-OSDCloud -ImageFileItem "\windows 11 - 24H2 - NLD\install.wim" -OSImageIndex 3